# core/vm: improved stack swap performance