This PR adds the methods `Stack.swap1..16()` that are ~25% faster than `Stack.swap(1..16)`. The improved performance originates from less slice bound checks of the assembly generated by the Go compiler. While `swap(1..16)` does two bound checks, `swap1..16()` only requires a single bound check.

For a detailed peek into the assembly generated for the relevant code see: https://go.godbolt.org/z/vrMfoYGzs

## Benchmarks
```
goos: linux
goarch: amd64
pkg: github.com/ethereum/go-ethereum/core/vm
cpu: 11th Gen Intel(R) Core(TM) i7-1185G7 @ 3.00GHz
            │   swap(1)   │               swap1()               │
            │   sec/op    │   sec/op     vs base                │
StackSwap-8   3.440n ± 1%   2.492n ± 1%  -27.57% (p=0.000 n=20)
```

```
goos: linux
goarch: amd64
pkg: github.com/ethereum/go-ethereum/core/vm/runtime
cpu: 11th Gen Intel(R) Core(TM) i7-1185G7 @ 3.00GHz
                │ stack_old.bench │           stack_new.bench           │
                │     sec/op      │   sec/op     vs base                │
EVM_SWAP1/10k-8       87.15µ ± 1%   77.84µ ± 1%  -10.68% (p=0.000 n=20)
```