@lmittmann 2024-08-01
> LGTM, but can't we also get rid of `swap(n int)` now?

I can drop it. If we keep it, we have the comparison in the benchmark. drop or keep?