@namiloh 2024-08-01
The benchmark is good for motivating the PR, but I don't think we need it long-term. Alternatively, add a bench as first commit, and then make the change. Then we can have before/after, drop the swap(n) and keep the bench