@lmittmann 2024-08-01
okay, I dropped `swap(n)` and the `stack_test.go` benchmark. I kept the benchmark in `runtime/` though.