@holiman 2024-08-06
On my machine, running with `--count 20` I can almost see a difference, but not sufficient for `benchstat` to think it is significant. 
```
goos: linux
goarch: amd64
pkg: github.com/ethereum/go-ethereum/core/vm/runtime
cpu: 12th Gen Intel(R) Core(TM) i7-1270P
                │ swap.before  │          swap.after           │
                │    sec/op    │   sec/op     vs base          │
EVM_SWAP1/10k-8   177.3µ ± 10%   168.5µ ± 6%  ~ (p=0.211 n=20)
```

And this is with a contract that is purely `swap1`, 10K operations. So I don't know. The overall gains here must be extremely slim, if even on such a non-realistic contract doesn't improve markedly. 

I think it's fine / safe to merge this PR. I'll leave it for triage to decide if we want to keep the original, slightly smaller code, or the new, slightly less complex and somewhat more LOC.