@lmittmann 2024-08-06
I am a bit skeptical about your benchmark results.
* You have the better CPU, yet your measurements are >100% slower (https://www.cpubenchmark.net/compare/4720vs3793/Intel-i7-1270P-vs-Intel-i7-1185G7)
* your measurements have a very high variability, which is often a sign for something in the background occupying the CPU as well. The high variability is also the reason benchstat does not consider the difference to be significant